==> before_everything_there_were_notebooks/01-Create the raw development and production datasets.ipynb <==
# 01 - Create the raw `development` and `production` dataseets

__Goal__:
1. Read the project's raw dataset: `"weather_dataset_raw.csv"`;
2. Sort it according to the `Timestamp` variable (remove the `UTC offsets`);
3. Split it into two datasets:
  - the  development dataset, spanning the years `2006-2010`
  - the production dataset, spanning  the years `2011-2016`;
4. Save the `developement` and `production` raw datasets, as `"weather_dataset_raw_develpment.csv"` and `"weather_dataset_raw_production.csv"` respectively.

### Import

import pandas as pd
from pathlib import Path

data_dir = Path('../data')

# 1. Read the raw dataset

# import dataset
df = pd.read_csv(data_dir/'weather_dataset_raw.csv')

df.info()

df.head()

df.tail()

# 2. Sort the raw dataset according to the `Timestep` variable

### A. Convert `Timestamp` variable's type to `datetime` and remove `UTC offsets`

df['Timestamp'] = pd.to_datetime(df['Timestamp'], utc=True)

### B. Sort the raw dataset

df.sort_values(by='Timestamp', inplace=True)

df.head()

df.tail()

# 3. Split the raw dataset

### A. The `development` dataset

development_period = (df["Timestamp"] >= "2006") & (df["Timestamp"] < "2011")
development_df = df[development_period]
development_df.info()

print(f'Number of raws of "development_df": {len(development_df)}')

development_df.head()

development_df.tail()

### B. The `production` dataset

production_period = df["Timestamp"] >= "2011"
production_df = df[production_period]
production_df.info()

production_df.head()

production_df.tail()

# 4. Save the `development` and `production` raw datasets

development_df.to_csv(data_dir/'weather_dataset_raw_development.csv', index=False)

production_df.to_csv(data_dir/'weather_dataset_raw_production.csv', index=False)
