==> before_everything_there_were_notebooks/02a-Preprocess the development  dataset - Timestamp as index.ipynb <==
# 02a - Preprocess the `development dataset` - `Timestamp` as `index`

__Goal__:

1. Read the `development` dataset `weather_dataset_raw_development.csv`, and check that it is sorted or sortit according to the `Timestamp` variable;
2. Preprocess the `Timestamp` variable, aka:
 - Remove rows with `Timestamp` duplicates;
 - Set the `Timestamp` variable as the `index` of the dataset;
 - Add empty rows (filled  with `Nan`s) to the dataset when a `Timestamp` is missing;
3. Save the `development` dataset as `weather_dataset_raw_development_timestamp.pkl`.

### Import

import pandas as pd
from pathlib import Path

data_dir = Path('../data')

# 1. Read the `development` dataset, and  check that it is sorted or sort it

## A. Read the `development` dataset

df = pd.read_csv(data_dir / 'weather_dataset_raw_development.csv')
df.info()

## B. Convert the `Timestamp` variable into a `datetime` without `UTC offsets`

df['Timestamp'] = pd.to_datetime(df['Timestamp'], utc=True)

type(df['Timestamp'][0])

## C. Check that the dataset is sorted according to the `Timestamp` variable

df.sort_values(by='Timestamp', inplace=True)

df.head()

df.tail()

# 2. Preprocess the `Timestamp` variable

## A. Remove `Timestamp` duplicates

df = df.drop_duplicates(subset=["Timestamp"], keep="last")
df.info() # 24 duplicates removed

df["Timestamp"].diff()

## B. Set the `Timestamp` variable as the `index` of the dataset

df.set_index('Timestamp', inplace=True)

## C. Add empty rows to the dataset when a `Timestamp` is missing

df_min_timestamp = df.index.min()
df_max_timestamp = df.index.max()
print(f'Minimum index of "df": {df_min_timestamp} \nMaximum index of "df": {df_max_timestamp}')

regular_timestamp_range = pd.date_range(start=df_min_timestamp, end=df_max_timestamp,freq='H')
print(f"Length of the dataframe `df`: {len(df)}\nLength of the datetime index `regular_timestamp_range`: {len(regular_timestamp_range)}")
if len(df) == len(regular_timestamp_range):
    print("\nNo timestamp was missing in the dataframe.")

regular_timestamp_df = df.reindex(regular_timestamp_range, copy=True)
regular_timestamp_df.info()

# 3. Save the `development` dataset as `weather_dataset_raw_development.pkl`

regular_timestamp_df.to_pickle(data_dir/'weather_dataset_raw_development_timestamp.pkl')
