==> before_everything_there_were_notebooks/02b-Preprocess the development dataset - Columns.ipynb <==
# 02b-Preprocess the `development` dataset - Columns

__Goal__:

1. Read the dataset `weather_dataset_raw_development_timestamp.pkl`
2. Preprocess the columns
   - Rename the columns
   - Drop the columns `S_No`, `Location`, `Apparent_temperature`
   - Refactor the column `Weather`
3. Statistical analysis of the numerical variables
4. Statistical analysis of the categorical variable `Weather`
5. Save the  preprocessed data as `weather_dataset_raw_development_columns.pkl`, and remove `weather_dataset_raw_development_timestamp.pkl`.

### Import

import os

import numpy as np
import pandas as pd
from IPython.display import display
from pathlib import Path

data_dir = Path('../data')

### Utilities

def statistical_analysis_of_numerical_variables(df):
    
    # Print the numerical variables
    numerical_variables = list(df.select_dtypes(include= "number").columns)
    print('Numerical variables of "df":')
    print("-"*len('Numerical variables of "df":')+"\n")
    print(', '.join(numerical_variables)+"\n")

    # Print the number of NaN's per numerical variables
    print("Number of NaN's per numerical variables:")
    print("-"*len("Number of NaN's per numerical variables:")+"\n")
    display(df[numerical_variables].isnull().sum())

    # Display the statistics of numerical variables
    print("\nStatistics of numerical variables:")
    print("-"*len("Statistics of numerical variables:")+"\n")
    display(df[numerical_variables].describe())

    # Display the correlation matrix of numerical variables
    print("\nCorrelation matrix of numerical variables:")
    print("-"*len("Correlation matrix of numerical variables:")+"\n")
    display(df[numerical_variables].corr('pearson'))

# 1. Read the  dataset

df = pd.read_pickle(data_dir/'weather_dataset_raw_development_timestamp.pkl')
df.info()

df.head()

# 2. Preprocess the columns

## A. Rename the columns

df.rename(columns={"Temperature_C": "Temperature", 
                   "Apparent_Temperature_C": "Apparent_temperature",
                   "Wind_speed_kmph": "Wind_speed",
                   "Wind_bearing_degrees": "Wind_bearing",
                   "Visibility_km": "Visibility",
                   "Pressure_millibars": "Pressure",
                   "Weather_conditions": "Weather"}, inplace=True)

df.head()

## B. Drop the columns `S_No`, `Location`, and `Apparent_temperature`

#### `S_No`

len(df["S_No"].unique())

As `S_No` (Serial number) seems to be an incremental index, we discard it.

df.drop(['S_No'], axis=1, inplace=True)

#### `Location`

df["Location"].value_counts()

df.drop(['Location'], axis=1, inplace=True)

#### `Apparent_temperature`

df[["Temperature", "Apparent_temperature"]].corr('pearson')

df.drop(['Apparent_temperature'], axis=1, inplace=True)

## C. Refactor the column `Weather`

df["Weather"].value_counts(dropna=False)

df["Weather"].replace({"snow": "no_rain", "clear": "no_rain"}, inplace=True)

df["Weather"].value_counts(dropna=False)

df["Weather"] = df["Weather"].map({'rain': 0, 'no_rain': 1})

df["Weather"].value_counts(dropna=False)

# 3. Statistical analysis of `numerical_variables`

statistical_analysis_of_numerical_variables(df)

# 4. Statistical analysis of `Weather`

df['Weather'].value_counts(dropna=False) 

# 5. Save the preprocessed dataset, and remove the previous one

df.to_pickle(data_dir/'weather_dataset_raw_development_columns.pkl')

os.remove(data_dir/'weather_dataset_raw_development_timestamp.pkl')
