==> before_everything_there_were_notebooks/02e-Preprocess the development dataset - Target column FutureWeather.ipynb <==
# 02e - Preprocess the `development` dataset - The target column `Future_Weather`

__Goal__: Create the target column `Future_Weather`

1. Read the clean dataset `clean_weather_dataset.pkl`
2. Add the target column `Future_Weather`
3. Save the dataset as `clean_weather_dataset_with_target.pkl`, and remove `clean_weather_dataset.pkl`.

### Import

import os
import pandas as pd
from pathlib  import  Path

data_dir = Path('../data')

# 1. Read the clean dataset

df = pd.read_pickle(data_dir/'clean_weather_dataset.pkl')
df.head()

df.tail()

# 2. Add the target column `Future_Weather`

df['Future_Weather'] = df["Weather"].shift(4, axis = 0) 
df.head()

df.tail()

### Percentage rain

print("Percentage rain", df['Future_Weather'].mean())

# 3. Save the dataset as `clean_weather_dataset_with_target.pkl`

df.to_pickle(data_dir/'clean_weather_dataset_with_target.pkl')

os.remove(data_dir/'clean_weather_dataset.pkl')
