==> before_everything_there_were_notebooks/07-Deepchecks_for_data_and_model_validations.ipynb <==
# 07 - `Deepchecks` for `Data validation` and `Model validation`

__Goal__:
1. Data validation with an `html` output
2. Data validation with a `boolean` output
3. Model validation with an `html` output
4. Model validation with a `boolean` output

### Import

%load_ext autoreload
%autoreload 2
import joblib
import os
import shutil

import pandas as pd
from pathlib import Path

from deepchecks.tabular import Dataset as DeepChecksDataset
from deepchecks.tabular.suites import data_integrity
from deepchecks.tabular.suites import model_evaluation

from weather.data.prep_datasets import (
    Dataset,
    transform_dataset_and_create_target,
    prepare_binary_classification_tabular_data,
)
from weather.transformers.skl_transformer_makers import (
    make_dataset_ingestion_transformer,
    make_predictors_feature_engineering_transformer,
    make_remove_horizonless_rows_transformer,
    make_target_creation_transformer,
)
from weather.pipelines.definitions import (
    feature_names,
    metric,
    oldnames_newnames_dict,
    target_choice,
)

### Directory paths

data_dir = Path('../data')
deepchecks_dir = Path('./deepchecks')
deepchecks_dir.mkdir(exist_ok=True)

models_dir = Path.cwd().parent / "models"
models_dir.mkdir(exist_ok=True)
model_subdir  = models_dir / "random_forest"
model_subdir.mkdir(exist_ok=True)

### Utilities

def validate_ingested_data(ingested_df, feature_names, target_choice):
    """Run the data integrity suite on `ingested_df`. Return True if all tests pass, False otherwise."""
    # Populate Dataset parameters
    features = feature_names.numerical + feature_names.categorical + [target_choice.input_name]
    cat_features = [target_choice.input_name]
    # Convert ingested_df into a deepchecks Dataset instance
    ds = DeepChecksDataset(ingested_df, features = features, cat_features = cat_features)
    # Run integrity suite 
    integrity_suite = data_integrity()
    results = integrity_suite.run(ds)
    return results.passed()

def validate_model(dataset, trained_model, trained_predictors_feature_engineering_transformer, excluded_check=5):
    """Run the validation suite minus `WeekSegmentPerformance` on `dataset`. Return True if all tests pass, False otherwise."""
    # Populate train_ds
    X_train = trained_predictors_feature_engineering_transformer.transform(dataset.train_x)
    y_train = dataset.train_y
    train_ds = DeepChecksDataset(X_train, label=y_train, cat_features=[])
    # Populate test_ds
    X_test = trained_predictors_feature_engineering_transformer.transform(dataset.test_x)
    y_test = dataset.test_y
    test_ds = DeepChecksDataset(X_test, label=y_test, cat_features=[])
    # Run model validation suite 
    evaluation_suite = model_evaluation()
    results = evaluation_suite.remove(excluded_check).run(train_ds, test_ds, trained_model)
    return results.passed()

# 1. Data validation with an `html` output

### Load raw data

csv_file_name = 'weather_dataset_raw_development.csv'
df = pd.read_csv(data_dir / csv_file_name)
df.info()

###  Ingest

dataset_ingestion_transformer = make_dataset_ingestion_transformer(target_choice, oldnames_newnames_dict)
ingested_df = dataset_ingestion_transformer.transform(df)
ingested_df.head(1)

### Transform into a `deepchecks.Dataset` object

features = ["Temperature", "Humidity", "Wind_speed",  "Wind_bearing", "Visibility", "Pressure", "Weather"]
cat_features = ["Weather"] # must be a subset of features

ds_name="ingested_data" 
ds = DeepChecksDataset(
    ingested_df,
    features=features,
    cat_features=cat_features,
)

### Check data integrity suite

integrity_suite = data_integrity()
integrity_suite

### Set results directory

deepchecks_subdir = deepchecks_dir / ds_name
deepchecks_subdir.mkdir(exist_ok=True)
print(deepchecks_subdir)

### Data Validation

results = integrity_suite.run(ds)

results_attributes = "\n".join([attributes for attributes in dir(results) if not attributes.startswith("_")])
print(f"Attributes of `results`:\n\n{results_attributes}")

### Export results

#results.save_as_html("integrity_suite_output.html")
#shutil.move("integrity_suite_output.html", deepchecks_subdir)

# 2. Data validation with a boolean output

### Load raw data and ingest it

csv_file_name = 'weather_dataset_raw_development.csv'
df = pd.read_csv(data_dir / csv_file_name)
dataset_ingestion_transformer = make_dataset_ingestion_transformer(target_choice, oldnames_newnames_dict)
ingested_df = dataset_ingestion_transformer.transform(df)
ingested_df.head(1)

### Data Validation

data_integrity_suite_passed = validate_ingested_data(ingested_df, feature_names, target_choice)
print(f"The data integrity suite passed: {data_integrity_suite_passed}")

# 3. Model validation with an `html` output

### Load raw data, ingest, transform and split it

# Read 
csv_file_name = 'weather_dataset_raw_development.csv'
df = pd.read_csv(data_dir / csv_file_name)

# Ingest and transform
remove_horizonless_rows_transformer = make_remove_horizonless_rows_transformer(target_choice)
target_creation_transformer = make_target_creation_transformer(target_choice)
transformed_data, created_target = transform_dataset_and_create_target(
    df,   
    dataset_ingestion_transformer,
    remove_horizonless_rows_transformer,
    target_creation_transformer,
)

# Split
dataset = prepare_binary_classification_tabular_data(
    transformed_data,
    created_target,
)

### Load trained model and fitted transformer

trained_model = joblib.load(model_subdir / "model.pkl")
trained_predictors_feature_engineering_transformer = joblib.load(model_subdir / "predictors_feature_eng_pipeline.pkl")

### Populate  `train_ds` and `train_val`

# Populate train_ds
X_train = trained_predictors_feature_engineering_transformer.transform(dataset.train_x)
y_train = dataset.train_y
train_ds = DeepChecksDataset(X_train, label=y_train, cat_features=[])

# Populate test_ds
X_test = trained_predictors_feature_engineering_transformer.transform(dataset.test_x)
y_test = dataset.test_y
test_ds = DeepChecksDataset(X_test, label=y_test, cat_features=[])

### Check model evaluation suite

evaluation_suite = model_evaluation()
evaluation_suite

### Model Validation

results = evaluation_suite.run(train_ds, test_ds, trained_model)

results_attributes = "\n".join([attributes for attributes in dir(results) if not attributes.startswith("_")])
print(f"Attributes of `results`:\n\n{results_attributes}")

results.get_not_passed_checks()

# 4. Model validation with a `boolean` output

### Load raw data, ingest, transform and split it

# Read 
csv_file_name = 'weather_dataset_raw_development.csv'
df = pd.read_csv(data_dir / csv_file_name)

# Ingest and transform
remove_horizonless_rows_transformer = make_remove_horizonless_rows_transformer(target_choice)
target_creation_transformer = make_target_creation_transformer(target_choice)
transformed_data, created_target = transform_dataset_and_create_target(
    df,   
    dataset_ingestion_transformer,
    remove_horizonless_rows_transformer,
    target_creation_transformer,
)

# Split
dataset = prepare_binary_classification_tabular_data(
    transformed_data,
    created_target,
)

### Load trained model and fitted transformer

trained_model = joblib.load(model_subdir / "model.pkl")
trained_predictors_feature_engineering_transformer = joblib.load(model_subdir / "predictors_feature_eng_pipeline.pkl")

### Model Validation

model_validation_suite_passed = validate_model(dataset, trained_model, trained_predictors_feature_engineering_transformer)
print(f"The model validation suite minus 'WeekSegmentPerformance' passed: {model_validation_suite_passed}")
