==> before_everything_there_were_notebooks/08-RawDataWithoutNaNs.ipynb <==
# 6c - Raw data without NaNs

### Import

import pandas as pd
import warnings 
warnings.filterwarnings('ignore')
from pathlib import Path

### Set the directory paths

data_dir =  Path.cwd().parent / "data"

### Read raw data

df = pd.read_csv(data_dir / 'weather_dataset_raw.csv')
len(df)

### Cast `Timestamp` into `datetime`

df['Timestamp'] = pd.to_datetime(df['Timestamp'], utc=True)

df.sort_values(by='Timestamp', inplace=True)
len(df)

df = df.drop_duplicates(subset=["Timestamp"], keep="last")
len(df) # 24 duplicates removed 

df.set_index('Timestamp', inplace=True)

df_min_timestamp = df.index.min()
df_max_timestamp = df.index.max()
print(f'Minimum index of "df": {df_min_timestamp} \nMaximum index of "df": {df_max_timestamp}')

regular_timestamp_range = pd.date_range(start=df_min_timestamp, end=df_max_timestamp,freq='H')
print(f"Length of the dataframe `df`: {len(df)}\nLength of the datetime index `regular_timestamp_range`: {len(regular_timestamp_range)}")
diff = len(regular_timestamp_range) - len(df)
if diff == 0:
    print("\nNo timestamp was missing in the dataframe.")
else:
    print(f"\n{diff} timestamp(s) were missing in the dataframe.")

df = df.reindex(regular_timestamp_range, copy=True)
len(df)

df["Timestamp"] = df.index

df.reset_index(drop=True, inplace=True)
df.head(1)

### Remove `NaNs`

df.isna().sum().sum()

df[:1] = df.bfill()[:1]

df.ffill(inplace=True)

df.isna().sum().sum()

### Split data into `dev` and `prod`

dev_period = (df["Timestamp"] >= "2006") & (df["Timestamp"] < "2011")
dev_df = df[dev_period]
dev_df
len(dev_df)

dev_df.head(1)

prod_period = df["Timestamp"] >= "2011"
prod_df = df[prod_period]
len(prod_df)

prod_df.head(1)

###  Store `dev` and `prod` data as `csv` files

dev_df.to_csv(data_dir/'weather_dataset_raw_development.csv', index=False)

prod_df.to_csv(data_dir/'weather_dataset_raw_production.csv', index=False)
